restructured navbar files
