+++ client/src/Components/NavBar/Public.js
@@ -1,43 +1,31 @@
-import React, { useState } from 'react';
+import React from 'react';
 import { useHistory } from 'react-router-dom';
 import { Menu } from 'antd';
 
-import '../Layout/style.css';
+import '../../Pages/Layout/style.css';
 
 const PublicMenu = () => {
-  const [current, setCurrent] = useState('home');
   const history = useHistory();
 
-  const handleMenu = (e) => {
-    setCurrent(e.key);
-  };
-
   return (
-    <div className="menu-public">
-      <Menu
-        className="nav-menu"
-        onClick={handleMenu}
-        selectedKeys={[current]}
-        mode="horizontal"
+    <>
+      <Menu.Item
+        onClick={() => {
+          history.push('/');
+        }}
+        key="home"
       >
-        <Menu.Item
-          onClick={() => {
-            history.push('/');
-          }}
-          key="home"
-        >
-          Home
-        </Menu.Item>
-        <Menu.Item
-          onClick={() => {
-            history.push('/about-us');
-          }}
-          key="About us"
-        >
-          About Us
-        </Menu.Item>
-      </Menu>
-    </div>
+        Home
+      </Menu.Item>
+      <Menu.Item
+        onClick={() => {
+          history.push('/about-us');
+        }}
+        key="About us"
+      >
+        About Us
+      </Menu.Item>
+    </>
   );
 };
