+++ client/src/Components/NavBar/User.js
@@ -1,89 +1,36 @@
-import React, { useState, useContext } from 'react';
-import PropTypes from 'prop-types';
+import React, { useContext } from 'react';
 import { useHistory } from 'react-router-dom';
-import { Menu, Spin } from 'antd';
+import { Menu } from 'antd';
 import AuthProvider from '../../Context/Authentication';
 
-import UserInfo from '../UserInfo';
-
-const UserMenu = ({ handleLogout, notifications, errMsg, isLoading }) => {
-  const [current, setCurrent] = useState('home');
+const UserMenu = () => {
   const { userData } = useContext(AuthProvider);
-  const { name, avatar } = userData;
+  const { role } = userData;
   const history = useHistory();
 
-  const handleMenu = (e) => {
-    setCurrent(e.key);
-  };
-
   return (
-    <div className="customer-container">
-      <Menu
-        className="nav-menu"
-        onClick={handleMenu}
-        selectedKeys={[current]}
-        mode="horizontal"
-      >
+    <>
+      {role === 'user' ? (
         <Menu.Item
           onClick={() => {
             history.push('/');
           }}
-          key="home"
-        >
-          Home
-        </Menu.Item>
-        <Menu.Item
-          onClick={() => {
-            history.push('/orders');
-          }}
           key="Orders"
         >
-          Orders
+          Order
         </Menu.Item>
+      ) : (
         <Menu.Item
           onClick={() => {
-            history.push('/about-us');
+            history.push('/provider/dashboard');
           }}
-          key="About us"
+          key="Dashboard"
         >
-          About Us
+          Dashboard
         </Menu.Item>
-      </Menu>
-      {errMsg ? (
-        <p className="error-msg">something went wrong</p>
-      ) : (
-        <>
-          {isLoading ? (
-            <Spin />
-          ) : (
-            <UserInfo
-              loggedUserInfo="logged-user-info"
-              userNameClass="user-name"
-              handleLogout={handleLogout}
-              shape="circle"
-              userPic={avatar}
-              userName={name}
-              notifications={notifications}
-            />
-          )}
-        </>
       )}
-    </div>
+    </>
   );
-};
-
-UserMenu.propTypes = {
-  notifications: PropTypes.arrayOf(
-    PropTypes.shape({
-      id: PropTypes.number,
-      user_id: PropTypes.number,
-      decription: PropTypes.string,
-      created_at: PropTypes.string,
-    })
-  ),
-  handleLogout: PropTypes.func,
-  errMsg: PropTypes.string,
-  isLoading: PropTypes.bool,
 };
 
 export default UserMenu;
